requirements.txt
tensorflow
numpy
pandas
scikit-learn
opencv-python-headless
matplotlib
seaborn
Flask
Pillow
gunicorn